Afegida la lluita entre unitats, així com el moviment pel mapa limitat al moviment total de la unitat per torn.

+++ Assets/ScriptableObjects/UnitStats.cs
@@ -7,7 +7,7 @@ public class UnitStats : ScriptableObject
 {
     public int Attack;
     public int Defense;
-    public int Movement;
+    public float Movement;
     public int Health;
 
     public enum UnitType { WARRIOR, SETTLER, SIZE };
